// pkgs/akanjs/signal/types.ts
interface InitOption {
  serverMode?: "federation" | "batch" | "all";
  operationMode?: ("cloud" | "edge" | "local" | (string & {}))[];
  enabled?: boolean;
}

interface TimerOption {
  serverMode?: "federation" | "batch" | "all";
  operationMode?: ("cloud" | "edge" | "local" | (string & {}))[];
  lock?: boolean;
  enabled?: boolean;
}

export interface SignalOption<Response = any, Nullable extends boolean = false, _Key = keyof UnCls<Response>>
  extends InitOption,
    TimerOption {
  nullable?: Nullable;
  name?: string;
  default?: boolean;
  path?: string;
  serverMode?: "federation" | "batch" | "all";
  timeout?: number;
  partial?: _Key[] | readonly _Key[];
  cache?: number;
  guards?: GuardCls[];
  middlewares?: MiddlewareCls[];
  prefix?: false | string;
  globalPrefix?: false;

  // * ==================== Schedule ==================== * //
  scheduleType?: "init" | "destroy" | "cron" | "interval" | "timeout";
  scheduleCron?: string;
  scheduleTime?: number;
  lock?: boolean;
  enabled?: boolean;
  // * ==================== Schedule ==================== * //
}